pyproject.toml
[project]
name = "mcp-core"
version = "1.0.0"
description = "MCP记忆控制机制 - 生产级实现"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "MCP Development Team", email = "dev@mcp-project.com"}
]

dependencies = [
    "fastapi>=0.108.0",
    "uvicorn[standard]>=0.25.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.23",
    "alembic>=1.13.0",
    "pymysql>=1.1.0",
    "cryptography>=41.0.0",  # 用于PyMySQL的加密连接
    "redis>=5.0.1",
    "pymilvus>=2.3.4",
    "sentence-transformers>=2.2.2",
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "pyyaml>=6.0.1",
    "httpx>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "prometheus-client>=0.19.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "locust>=2.20.0",
    "black>=23.12.0",
    "ruff>=0.1.9",
    "mypy>=1.7.1",
    "ipython>=8.18.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src/mcp_core --cov-report=html --cov-report=term-missing"
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ['py310']
exclude = '''
/(
    \.git
  | \.venv
  | \.pytest_cache
  | __pycache__
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B"]
ignore = ["E501", "B008"]
exclude = [".venv", "__pycache__", "build", "dist"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
